+++ src/app/(public)/mate/selectOptionArray.ts
@@ -1,34 +1,24 @@
 export const characteristicsArr = ["선택", "온순함", "활발함", "소심함", "적극적", "외향적", "내향적", "낯가림"];
 
 export const gender = [
-  {key: "전체", label: "전체"},
-  {key: "남", label: "남"},
-  {key: "여", label: "여"},
+  { key: "전체", label: "전체" },
+  { key: "남", label: "남" },
+  { key: "여", label: "여" }
 ];
 
 export const age = [
-  {key: "전체", label: "전체"},
-  {key: "10대", label: "10대"},
-  {key: "20대", label: "20대"},
-  {key: "30대", label: "30대"},
-  {key: "40대", label: "40대"},
-  {key: "50대", label: "50대"},
-  {key: "60대", label: "60대"},
-  {key: "70대이상", label: "70대이상"},
+  { key: "전체", label: "전체" },
+  { key: "10대", label: "10대" },
+  { key: "20대", label: "20대" },
+  { key: "30대", label: "30대" },
+  { key: "40대", label: "40대" },
+  { key: "50대", label: "50대" },
+  { key: "60대", label: "60대" },
+  { key: "70대이상", label: "70대이상" }
 ];
 
-// export const position = [
-//   {key: "0", label: "0"},
-//   {key: "5", label: "5"},
-//   {key: "10", label: "10"},
-//   {key: "15", label: "15"},
-//   {key: "30", label: "30"},
-//   {key: "50", label: "50"},
-//   {key: "100", label: "100"},
-// ];
-
 export const regions = [
-  {key: "전체", label: "전체"},
+  { key: "전체", label: "전체" },
   { key: "서울", label: "서울" },
   { key: "인천", label: "인천" },
   { key: "경기", label: "경기" },
@@ -47,10 +37,8 @@ export const regions = [
   { key: "제주", label: "제주" }
 ];
 
-
 export const male_female = [
-  {key: "전체", label: "전체"},
-  {key: "male", label: "남"},
-  {key: "female", label: "여"},
+  { key: "전체", label: "전체" },
+  { key: "male", label: "남" },
+  { key: "female", label: "여" }
 ];
-
